22.11.21 작업사항(close #72) (#74)
* 초기 설정 (#2)

* chore: github 템플릿 추가

* chore: prettier 적용

* build: next 초기 설정

* build: emotion 설치

* build: axios 설치

* chore: 클라이언트 CI/CD

* chore: 실행 권한 부여


* chore: deploy.yml 삭제

* 1 공통 - 상단(GNB) #5 (#55)

* feat: global css 적용


* feat: GNB 컴포넌트 작성


* chore: strict 모드 해제, eslint 수정


* feat: Close #6


* feat: user 로그인 상태 mock api 구현


* feat: Close #15, #16


* fix: login status api 수정

로그인시 userName 추가


* feat: Close #14


* feat: GNB 적용



* 2. 문제 목록 (#8) (#56)

* feat: 2-1 head 에는 문제 번호, 제목, 정답 비율이 위치한다 (#24)


* feat: 2-2 문제 리스트를 20개씩 보여준다. (#30)


* feat: mock api 작성


* feat: 2-5 처음 들어왔을 때는 전체 문제 목록을 보여준다. (#34)


* feat: 2-4 리스트 행 클릭시 문제 상세 페이지로 이동한다. (#33)


* chore: list style 수정


* feat: 2-6 페이지네이션이 리스트 하단에 위치한다. (#35)



* build: deploy.sh 수정 (#58)

yan install 추가

* 4. 채점 현황 (#12) (#59)

* refactor: 일반화된 리스트 컴포넌트 작성


* refactor: List 컴포넌트 관련 하나의 폴더로 묶음


* refactor: 다시 내보내기를 이용한 import 경로 간소화

* feat: 리스트 아이템 포멧 기능 추가

* feat: 4. 채점 현황 (Close #12)



* Feat/#7 (#60)

* feat: Button 컴포넌트 작성


* feat: global css 수정

root div 꽉 차게 수정


* feat: 문제 상세 페이지 마크업 Close #28


* feat: 문제 상세 mock API 구현


* fix: problem detail 페이지 height 수정

control panel border 만큼 height 수정


* feat: global.d.ts 추가


* fix: io 형식 수정


* feat: 문제 container 작성

추후 수정 예정
- plain text를 mark down 으로 수정 예정


* build: monaco editor 추가


* fix: problem ioExample key 오류 수정


* feat: close #26

monaco code editor 추가


* feat: code state 끌어올리기


* fix: 파일명 수정

problem -> problems


* feat:  코드제출 mock api 생성


* feat: close #32

제출버튼 클릭시 api요청 후 성공하면 채점 현황 페이지로 이동



* 6. 문제 출제 (Close #9) (#61)

feat: svg 파일 추가



* Feat/#11 (close #11) (#62)

* feat: 풀이 상세 코드 mock api 생성


* feat: close #19, close #20

코드 에디터에 사용자의 풀이를 띄우고 수정 불가하게 함


* feat: 풀이 상세 mock api status 추가


* feat: close #17

풀이 status list 추가



* fix: pages 폴더 내부 .ts 파일 제거

* chore: 파일 끝 엔터 추가

* fix: page 내부 ts 파일 삭제


* 6. 문제 출제 (#9) [추가 변경사항] (#66)

* feat: mock 데이터 추가, API 수정


* feat: 문제 삭제 모달


* feat: 문제 공개 여부 Mock API 생성


* feat: 문제 공개여부 토글



* Feat/#13 (#67)

* feat: 테케 mock api 생성


* feat: button 컴포넌트 style 속성 추가

- default, cancel style 추가
- minWidth, minHeight style 추가


* feat: global에 tc interface 추가


* feat: close #25, close #29

- 입력칸 component 생성
- 입력과 출력을 입력칸 component로 작성
- 삭제 버튼 추가


* feat: close #27, close #38, close #39, close #40

- close #23
- 하단에 저장, 취소 버튼 추가
- 추가버튼 클릭시 입출력 row 생성
- custom hook 생성(useIOList())
- 테케 1개 반드시 존재 확인(useEffect로 testcase 0개면 바로 테스트케이스 하나 추가해주는 방식)
- 확인/취소 버튼 클릭시 출제 목록 페이지로 이동



* feat: close #46 close #47 close #48 close #49 close #50 close #51 close #52


* feat: close #53 close #54


* feat: 7. 문제 추가/편집(close #10)


* Feat/#7 (#68)

* build: toast ui viewer 설치


* feat: close #21

- 좌측 영역에 문제 설명이 나타남
- 마크다운 에디터 뷰어를 사용하여 구현


* refactor: 필요없는 export 제거


* feat: close #22

- 문제 제목 제외 스크롤 가능
- Refactor style
- tc 입출력 input -> textarea로 변경(여러줄 입력을 위해)



* close #7

* hotfix/#70 (#71)

* fix: 페이지 하단 여백 추가

* fix: 위로 포커스 이동

* fix: 문제 삭제 기능 수정

* fix: 문제 편집 페이지 제목 수정

* Refactor/#72 (#73)

* refactor: page style 분리


* refactor: 자주 쓰는 type global로 분리

- props 제외 분리


* refactor: axios 통일


* refactor: list component paginator 선택 옵션, 취소, 저장 버튼 바꾸기


* fix: 마크다운에디터 옵션 수정

- 기본 에디터 마크다운으로 수정
- 마크다운 옵션에서 task, image 제거


* refactor: 모달 컴포넌트 분리


* refactor: 필요없는 props 제거
