+++ client/pages/my-problem/tc/[id].tsx
@@ -1,37 +1,9 @@
-import { css } from '@emotion/react';
-import axios from 'axios';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
+import axiosInstance from '../../../axios';
 import Button from '../../../components/common/Button';
 import IOList, { useIOList } from '../../../components/common/IOList';
-
-const style = {
-  container: css`
-    margin: 64px 20%;
-  `,
-  title: css`
-    font-size: 32px;
-    font-weight: bold;
-    margin: 42px 12px;
-  `,
-  header: css`
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  `,
-  titleContainer: css`
-    display: flex;
-    align-items: center;
-  `,
-  problemTitle: css`
-    color: #3949ab;
-  `,
-  footer: css`
-    display: flex;
-    justify-content: flex-end;
-    padding: 10px 0px;
-  `,
-};
+import style from '../../../styles/style';
 
 function Testcase() {
   const router = useRouter();
@@ -44,7 +16,9 @@ function Testcase() {
 
     (async () => {
       try {
-        const result = await (await axios(`/api/problems/${id}/tc`)).data;
+        const result = await (
+          await axiosInstance(`/api/problems/${id}/tc`)
+        ).data;
         setTitle(result.title);
         setTestCase(result.data);
       } catch (err) {
@@ -65,7 +39,7 @@ function Testcase() {
           throw new Error('빈칸..');
         }
       });
-      await axios.post(`/api/problems/${id}/tc`, testcase);
+      await axiosInstance.post(`/api/problems/${id}/tc`, testcase);
       router.push('/my-problem');
     } catch (err) {
       console.error(err);
