+++ client/pages/status/[id].tsx
@@ -1,31 +1,11 @@
-import axios from 'axios';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import Button from '../../components/common/Button';
 import { css } from '@emotion/react';
 import Editor from '@monaco-editor/react';
 import StatusList from '../../components/status/StatusList';
-
-const style = {
-  container: css`
-    margin: 64px 20%;
-    height: 80vh;
-  `,
-  title: css`
-    font-size: 32px;
-    font-weight: bold;
-    margin: 42px 12px;
-  `,
-};
-
-interface StatusSummary {
-  id: number;
-  user: string;
-  title: string;
-  result: number;
-  time: string;
-  datetime: number;
-}
+import style from '../../styles/style';
+import axiosInstance from '../../axios';
 
 function StatusDetail() {
   const router = useRouter();
@@ -35,7 +15,9 @@ function StatusDetail() {
 
   useEffect(() => {
     (async () => {
-      const res = await (await axios.get(`/api/submissions/${id}`)).data;
+      const res = await (
+        await axiosInstance.get(`/api/submissions/${id}`)
+      ).data;
       setCode(res.code);
       setStatus(res.status);
     })();
@@ -49,15 +31,7 @@ function StatusDetail() {
       ) : (
         <>
           <StatusList status={status} />
-          <div
-            css={css`
-              height: 80%;
-              padding: 15px 0;
-              border-radius: 8px;
-              margin-top: 15px;
-              box-shadow: 0px 0px 1.5px 1.5px rgba(0, 0, 0, 0.15);
-            `}
-          >
+          <div css={style.codeBox}>
             <Editor
               defaultLanguage="python"
               defaultValue={code}
@@ -76,14 +50,7 @@ function StatusDetail() {
           </div>
         </>
       )}
-      <div
-        css={css`
-          display: flex;
-          justify-content: flex-end;
-          padding: 10px 1px;
-          width: 100%;
-        `}
-      >
+      <div css={style.footer}>
         <Button>복사</Button>
       </div>
     </div>
