+++ client/styles/index.ts
@@ -1,0 +1,2 @@
+export { default as style } from './style';
+export { default as modal } from './modal';
